Added filter module for MultiParticleDataCall. Added CPU line rendering to MoleculeCartoonRenderer. Changed Unlocker in VTKLegacyDataLoaderUnstructuredGrid. Some minor changes to remove warnings.
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein/trunk@1192 5e8c2511-817d-0410-85a1-fd227d6480d1

+++ BindingSiteCall.h
@@ -83,7 +83,7 @@ namespace protein {
             if( !this->bindingSites )
                 return 0;
             else
-                return this->bindingSites->Count();
+                return static_cast<unsigned int>(this->bindingSites->Count());
         }
         
         /**
